+++ src/components/Footer/Footer.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta } from '@storybook/react';
 
 import Footer, { FooterProps } from './Footer';
-import Subscribe from '../Subscribe/Subscribe';
 import { px } from '../../utils';
 
 import Youtube from '../../assets/youtube.svg?react';
@@ -9,7 +8,8 @@ import Instagram from '../../assets/instagram.svg?react';
 import Wechat from '../../assets/wechat.svg?react';
 import Spotify from '../../assets/spotify.svg?react';
 import Social from '../Social/Social';
-import { SubscriptionState } from '../Subscribe/types';
+import { Accordion, AccordionItem } from '../Accordion';
+import { Text, TextVariants } from '../Text';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta = {
@@ -19,25 +19,59 @@ const meta = {
 
 export default meta;
 
-const navigationComponent = (
+const aboutUsLinks = (
   <ul>
     <li>
-      <a>Locations</a>
+      <a href="/">Our History</a>
     </li>
     <li>
-      <a>Press</a>
+      <a href="/">Our Team</a>
     </li>
     <li>
-      <a>Careers</a>
+      <a href="/">Locations</a>
     </li>
     <li>
-      <a>Privacy policy</a>
+      <a href="/">Press</a>
     </li>
     <li>
-      <a>Cookie policy</a>
+      <a href="/">Careers</a>
     </li>
     <li>
-      <a>Modern Slavery Policy</a>
+      <a href="/">Site map</a>
+    </li>
+  </ul>
+);
+
+const ourServicesLinks = (
+  <ul>
+    <li>
+      <a href="/">How to Buy</a>
+    </li>
+    <li>
+      <a href="/">How to Sell</a>
+    </li>
+    <li>
+      <a href="/">Private Services</a>
+    </li>
+    <li>
+      <a href="/">Trusts, Estates & Valuations</a>
+    </li>
+    <li>
+      <a href="/">Fiduciary Services</a>
+    </li>
+  </ul>
+);
+
+const policyLinks = (
+  <ul>
+    <li>
+      <a href="/">Privacy Policy</a>
+    </li>
+    <li>
+      <a href="/">Cookie Policy</a>
+    </li>
+    <li>
+      <a href="/">Modern Day Slavery Policy</a>
     </li>
   </ul>
 );
@@ -68,28 +102,57 @@ const socialIcons = (
 );
 
 const leftComponent = (
-  <Subscribe
-    className={`${px}-footer__newsletter`}
-    title="Subscribe to Newsletter"
-    blurb="Receive exclusive content about our auctions, exhibitions, and special events."
-    buttonText="Sign Up"
-    element="form"
-    buttonProps={{
-      size: 'sm',
-    }}
-    onSubmit={(e: React.FormEvent<HTMLFormElement>) => {
-      const target = e.target as HTMLFormElement;
-      e.preventDefault();
+  <>
+    <div className={`${px}-footer-desktop`}>
+      <div>
+        <Text variant={TextVariants.heading4}>About Us</Text>
+        {aboutUsLinks}
+      </div>
 
-      const nativeData = new FormData(target);
-      const data = Object.fromEntries(nativeData.entries());
-      console.log('Form submitted for email -', data?.email);
+      <div>
+        <Text variant={TextVariants.heading4}>Our Services</Text>
+        {ourServicesLinks}
+      </div>
 
-      target.reset();
-    }}
-    subscriptionState={SubscriptionState.Default}
-    invalidText="Please enter a valid email address."
-  />
+      <div>
+        <Text variant={TextVariants.heading4}>Policies</Text>
+        {policyLinks}
+      </div>
+    </div>
+
+    <Accordion type="multiple" className={`${px}-footer-mobile`}>
+      <AccordionItem
+        isLocked={false}
+        variation=""
+        label="About Us"
+        key="accordion-key-about-us"
+        id="accordion-item-about-us"
+      >
+        {aboutUsLinks}
+      </AccordionItem>
+
+      <AccordionItem
+        isLocked={false}
+        variation=""
+        label="Our Services"
+        key="accordion-key-our-services"
+        id="accordion-item-our-services"
+      >
+        {ourServicesLinks}
+      </AccordionItem>
+
+      <AccordionItem
+        isLocked={false}
+        variation=""
+        label="Policies"
+        isLastItem={true}
+        key="accordion-key-policies"
+        id="accordion-item-policies"
+      >
+        {policyLinks}
+      </AccordionItem>
+    </Accordion>
+  </>
 );
 
 const rightComponent = <Social className={`${px}-footer__social`}>{socialIcons}</Social>;
@@ -102,5 +165,5 @@ export const Playground = (props: FooterProps) => (
 );
 
 Playground.args = {
-  navigationComponent,
+  // navigationComponent,
 };