+++ src/components/Subscribe/Subscribe.tsx
@@ -4,6 +4,7 @@ import { getCommonProps } from '../../utils';
 import Input from '../Input/Input';
 import Button, { ButtonProps } from '../Button/Button';
 import { SubscriptionState } from './types';
+import { ButtonVariants } from '../Button/types';
 
 export interface SubscribeProps extends React.HTMLAttributes<HTMLFormElement> {
   /**
@@ -51,6 +52,10 @@ export interface SubscribeProps extends React.HTMLAttributes<HTMLFormElement> {
    */
   successText?: string;
   /**
+   * Subscribe privacy text
+   */
+  privacyText?: string;
+  /**
    * Subscribe state for loading or error
    */
   subscriptionState?: SubscriptionState;
@@ -72,13 +77,14 @@ const Subscribe = ({
   buttonProps,
   className,
   element: Element = 'form',
-  inputLabelText = 'Email*',
+  inputLabelText = 'Enter Your Email Address To Sign Up*',
   inputPlaceholder = 'example@email.com',
-  title = 'Subscribe to Newsletter',
+  title = 'Never Miss A Moment',
   loadingText = 'Loading...',
   invalidText = '',
   errorText = 'An error occurred. Please try again.',
   successText,
+  privacyText = 'By signing up, you agree to receive email communications from Phillips.',
   subscriptionState = SubscriptionState.Default,
   ...props
 }: SubscribeProps) => {
@@ -117,9 +123,17 @@ const Subscribe = ({
         warnText={text}
         required
       />
-      <Button className={`${baseClassName}__button ${className}`} buttonType="secondary" type="submit" {...buttonProps}>
+      <Button
+        className={`${baseClassName}__button ${className}`}
+        variant={ButtonVariants.secondary}
+        size="sm"
+        type="submit"
+        {...buttonProps}
+      >
         {buttonText}
       </Button>
+
+      {privacyText ? <p className={`${baseClassName}__privacy`}>{privacyText}</p> : null}
     </Element>
   );
 };